refactor: split device-registry to separate file
