+++ debian/install
@@ -5,6 +5,7 @@ wb-alice-setting.conf /etc/
 50wb-alice /etc/wb-configs.d/
 wb-alice-client.py /usr/lib/wb-mqtt-alice/
 mqtt_topic.py /usr/lib/wb-mqtt-alice/
+device-registry.py /usr/lib/wb-mqtt-alice/
 wb-alice-config.py /usr/lib/wb-mqtt-alice/
 models.py /usr/lib/wb-mqtt-alice/
 fetch_url.py /usr/lib/wb-mqtt-alice/
